Revert "Event marker sliders"
This reverts commit c52cf070451de28c9ed998d900a82271bafec682.
