+++ Editor/CustomEditors/AnimationClipAssetEditor.cs
@@ -11,7 +11,6 @@ namespace DOTSAnimation.Editor
     public class AnimationClipAssetEditor : UnityEditor.Editor
     {
         public VisualTreeAsset EditorXml;
-        public VisualTreeAsset EventMarkerXml;
         private SingleClipPreview preview;
         private AnimationClipAsset ClipTarget => (AnimationClipAsset)target;
         
@@ -37,7 +36,7 @@ namespace DOTSAnimation.Editor
                 objField.RegisterValueChangedCallback(OnPreviewObjectChanged);
                 
                 var eventsEditorView = inspector.Q<AnimationEventsEditorView>();
-                eventsEditorView.Initialize(ClipTarget, serializedObject, EventMarkerXml);
+                eventsEditorView.Initialize(ClipTarget, serializedObject);
                 eventsEditorView.SampleTimeChanged += OnSampleTimeChanged;
             }
