--- Editor/CustomEditors/AnimationEventsPropertyDrawer.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 81959fe1e3264ee18649bf7f1168ea6d
-timeCreated: 1657764801
