+++ Editor/CustomEditors/SliderDragger.cs
@@ -39,11 +39,11 @@ namespace DOTSAnimation.Editor
         {
         }
 
-        public float PositionToNormalizedValue(float posX)
+        private float PositionToNormalizedValue(float posX)
         {
             return posX / (dragArea.layout.width - dragElement.layout.width);
         }
-        public float NormalizedValueToPosition(float n)
+        private float NormalizedValueToPosition(float n)
         {
             var halfWidth = dragElement.layout.width * 0.5f;
             return Mathf.Lerp(-halfWidth, dragArea.layout.width - halfWidth, n);
