+++ Editor/UI/AnimationEventsEditorView.uxml
@@ -8,7 +8,9 @@
                 <ui:VisualElement name="dragger-container" class="unity-base-slider__dragger">
                     <ui:VisualElement name="unity-dragger" picking-mode="Ignore" class="unity-base-slider__dragger" />
                 </ui:VisualElement>
-                <ui:VisualElement picking-mode="Ignore" name="container__event-markers" style="flex-shrink: 0; flex-grow: 1; width: 100%;" />
+                <ui:VisualElement picking-mode="Ignore" name="container__event-markers" style="flex-shrink: 0; flex-grow: 1; width: 100%;">
+                    <ui:VisualElement class="event-marker" />
+                </ui:VisualElement>
             </ui:VisualElement>
             <ui:VisualElement name="button-container">
                 <ui:Button display-tooltip-when-elided="true" name="button-add-event" class="button__add-event">
@@ -16,6 +18,9 @@
                     <ui:Label text="+" display-tooltip-when-elided="true" picking-mode="Ignore" class="label-plus" />
                 </ui:Button>
             </ui:VisualElement>
+        </ui:VisualElement>
+        <ui:VisualElement>
+            <uie:PropertyField name="property-events" />
         </ui:VisualElement>
     </DOTSAnimation.Editor.AnimationEventsEditorView>
 </ui:UXML>
