--- Editor/UI/event-marker.uxml.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 71e82bb7913aa774cb669fb6afbf8d7b
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 13804, guid: 0000000000000000e000000000000000, type: 0}
